Move UserInfo to the Models folder
