+++ lib/Models/user_info.dart
@@ -1,0 +1,14 @@
+class UserInfo {
+  String firstName;
+  String secondName;
+  String email;
+  String password;
+  String userName;
+  String mobile;
+  String gender;
+  String height;
+  String weight;
+
+  UserInfo(this.firstName, this.secondName, this.email, this.password,
+      this.userName, this.mobile, this.gender, this.height, this.weight);
+}
