+++ lib/Registration/signup_height.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-
 // TODO: Find null safe replacement library
 import 'package:flutter_circular_slider/flutter_circular_slider.dart'; // ignore: import_of_legacy_library_into_null_safe
-import 'package:redback_mobile_app/Registration/sign_up.dart';
+import 'package:redback_mobile_app/Models/user_info.dart';
 import 'package:redback_mobile_app/Registration/signup_weight.dart';
 
 // flutter run --no-sound-null-safety
